+++ app/routes/termliste/filters/translation-filter.ts
